--- app/helpers/deserts_helper.rb
@@ -1,2 +1,0 @@
-module DesertsHelper
-end
