+++ app/helpers/desserts_helper.rb
@@ -1,0 +1,2 @@
+module DessertsHelper
+end
